+++ ruler-models/src/commonMain/kotlin/com/spotify/ruler/models/DynamicFeature.kt
@@ -22,8 +22,8 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class DynamicFeature(
     override val name: String,
-    override val downloadSize: Long,
-    override val installSize: Long,
+    override val downloadSize: Double,
+    override val installSize: Double,
     override val files: List<AppFile>?,
     override val owner: String? = null,
 ) : FileContainer
